:art: Add product list style

+++ shop/models.py
@@ -4,7 +4,7 @@ class Article(models.Model):
     title = models.CharField(max_length=100)
     image = models.ImageField(upload_to='images/',)
     description = models.TextField()
-    price = models.IntegerField()
+    price = models.FloatField()
     quantity = models.IntegerField()
     def __str__(self):
         return self.title
